// src/interfaces/monitor.interface.ts
export interface IMonitorErrorMessage {
  [x: string]: string | undefined;
  notificationId?: string;
  name?: string;
  frequency?: string;
  alertThreshold?: string;
  url?: string;
  timeout?: string;
  redirects?: string;
  method?: string;
  headers?: string;
  body?: string;
  httpAuthMethod?: string;
  basicAuthUser?: string;
  basicAuthPass?: string;
  bearerToken?: string;
  contentType?: string;
  statusCode?: string;
  responseTime?: string;
  connection?: string;
  port?: string;
}